+++ src/index.ts
@@ -1,4 +1,23 @@
 import {margin} from './margin/margin';
 import {padding} from './padding/padding';
+import {border} from './border/border';
+import {borderTop} from './border/borderTop';
+import {borderBottom} from './border/borderBottom';
+import {borderLeft} from './border/borderLeft';
+import {borderRight} from './border/borderRight';
+import {borderColor} from './border/borderColor';
+import {borderStyle} from './border/borderStyle';
+import {borderWidth} from './border/borderWidth';
 
-export {margin, padding};
+export {
+  margin,
+  padding,
+  border,
+  borderBottom,
+  borderColor,
+  borderLeft,
+  borderRight,
+  borderStyle,
+  borderTop,
+  borderWidth,
+};