+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Added `margin` and `padding` shorthands
+
 ## v0.1.0 (August 19, 2023)
 
 - Initial release