+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to the library will be documented in this file.
 
 - Added `margin` and `padding` shorthands.
 - Added `borderWidth` shorthand.
+- Added `borderStyle` shorthand.
 
 ## v0.1.0 (August 19, 2023)