+++ CHANGELOG.md
@@ -4,11 +4,13 @@ All notable changes to the library will be documented in this file.
 
 ## vX.X.X (Month DD, YYYY)
 
+## v0.2.0 (August 19, 2023)
+
 - Added `margin` and `padding` shorthands.
 - Added `borderWidth` shorthand.
 - Added `borderStyle` shorthand.
 - Added `borderColor` shorthand.
-- Added `borderTop`, `borderBottom`, `borderRight` and `borderLeft` shorthand.
+- Added `borderTop`, `borderBottom`, `borderRight` and `borderLeft` shorthands.
 - Added `border` shorthand.
 
 ## v0.1.0 (August 19, 2023)